backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database ORM and migrations
sqlmodel==0.0.22
psycopg2-binary==2.9.10
alembic==1.13.3

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0  # Pin to <5.0.0 due to passlib compatibility issues
python-multipart==0.0.12

# Environment variables
python-dotenv==1.0.1

# Development dependencies
black==25.12.0
isort==7.0.0
